requirements_cu114.txt
numpy
pandas
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cu113
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.13.1+cu113
-f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.12.1
matplotlib==3.3.1
protobuf==3.20.1
opencv-python
onnx==1.12
onnxruntime-gpu==1.12
mxnet-cu112
scikit-image==0.19.2
insightface==0.2.1
requests==2.25.1
kornia==0.5.4
dill
wandb